app/requirements.txt
annotated-types==0.7.0 #
anyio==4.6.2.post1 #
certifi==2024.8.30 #
cffi==1.17.1 #
charset-normalizer==3.4.1 #
# click==8.1.8
colorama==0.4.6 #
cryptography==44.0.0 #
distro==1.9.0 #
exceptiongroup==1.2.2 #
flake8==7.1.1
h11==0.14.0 #
httpcore==1.0.7 #
httpx==0.27.2 #
idna==3.10 #
isort==5.13.2
jiter==0.7.1 #
lxml==5.3.0 #
# mccabe==0.7.0
openai==1.55.1
pdfminer.six==20231228 #
pillow==11.0.0 #
# pycodestyle==2.12.1
pycparser==2.22 #
pydantic==2.10.1 #
pydantic_core==2.27.1 #
PyMuPDF==1.25.2
# pyflakes==3.2.0
pypdfium2==4.30.1 #
python-docx==1.1.2
python-dotenv==1.0.1
python-telegram-bot==21.7
# requests==2.32.3
sniffio==1.3.1 #
# starlette==0.41.3
# tornado==6.4.2
tqdm==4.67.1 #
typing_extensions==4.12.2 #
# urllib3==2.3.0
